+++ flutter_local_notifications/android/src/main/java/com/dexterous/flutterlocalnotifications/models/NotificationDetails.java
@@ -3,9 +3,8 @@ package com.dexterous.flutterlocalnotifications.models;
 import android.graphics.Color;
 import android.os.Build.VERSION;
 import android.os.Build.VERSION_CODES;
-
 import androidx.annotation.Keep;
-
+import androidx.annotation.Nullable;
 import com.dexterous.flutterlocalnotifications.NotificationStyle;
 import com.dexterous.flutterlocalnotifications.RepeatInterval;
 import com.dexterous.flutterlocalnotifications.models.styles.BigPictureStyleInformation;
@@ -14,9 +13,9 @@ import com.dexterous.flutterlocalnotifications.models.styles.DefaultStyleInforma
 import com.dexterous.flutterlocalnotifications.models.styles.InboxStyleInformation;
 import com.dexterous.flutterlocalnotifications.models.styles.MessagingStyleInformation;
 import com.dexterous.flutterlocalnotifications.models.styles.StyleInformation;
-
 import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 @Keep
@@ -118,8 +117,10 @@ public class NotificationDetails implements Serializable {
   private static final String FULL_SCREEN_INTENT = "fullScreenIntent";
   private static final String SHORTCUT_ID = "shortcutId";
   private static final String SUB_TEXT = "subText";
+  private static final String ACTIONS = "actions";
   private static final String COLORIZED = "colorized";
   private static final String NUMBER = "number";
+  private static final String AUDIO_ATTRIBUTES_USAGE = "audioAttributesUsage";
 
   public Integer id;
   public String title;
@@ -178,9 +179,11 @@ public class NotificationDetails implements Serializable {
   public Boolean fullScreenIntent;
   public String shortcutId;
   public String subText;
+  public @Nullable List<NotificationAction> actions;
   public String tag;
   public Boolean colorized;
   public Integer number;
+  public Integer audioAttributesUsage;
 
   // Note: this is set on the Android to save details about the icon that should be used when
   // re-hydrating scheduled notifications when a device has been restarted.
@@ -268,6 +271,21 @@ public class NotificationDetails implements Serializable {
       notificationDetails.tag = (String) platformChannelSpecifics.get(TAG);
       notificationDetails.colorized = (Boolean) platformChannelSpecifics.get(COLORIZED);
       notificationDetails.number = (Integer) platformChannelSpecifics.get(NUMBER);
+      notificationDetails.audioAttributesUsage =
+          (Integer) platformChannelSpecifics.get(AUDIO_ATTRIBUTES_USAGE);
+
+      if (platformChannelSpecifics.containsKey(ACTIONS)) {
+        @SuppressWarnings("unchecked")
+        List<Map<String, Object>> inputActions =
+            (List<Map<String, Object>>) platformChannelSpecifics.get(ACTIONS);
+        if (inputActions != null && !inputActions.isEmpty()) {
+          notificationDetails.actions = new ArrayList<>();
+          for (Map<String, Object> input : inputActions) {
+            final NotificationAction action = new NotificationAction(input);
+            notificationDetails.actions.add(action);
+          }
+        }
+      }
     }
   }