+++ flutter_local_notifications/ios/Classes/FlutterLocalNotificationsPlugin.h
@@ -2,4 +2,5 @@
 #import <UserNotifications/UserNotifications.h>
 
 @interface FlutterLocalNotificationsPlugin : NSObject <FlutterPlugin>
++ (void)setRegisterPlugins:(FlutterPluginRegistrantCallback *)callback;
 @end