+++ flutter_local_notifications_linux/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.0.0]
+
+* **Breaking change** The linux notification categories defined by `LinuxNotificationCategory` no longer has factory constructors but has static constant fields instead to make the semantics more similar to access enum values
+* Updated minimum Flutter version to 2.8 to align with the minimum Dart SDK version of 2.1.5 required by the `dbus` package
+
+
 ## [0.5.1]
 
 * Fixes issue [1656](https://github.com/MaikuB/flutter_local_notifications/issues/1656) where a version constraint issue occurred with the `path` package. The version constraint has been lowered to resolve the issue
@@ -10,6 +16,23 @@
 
 * Added support for icons to be specified via file path. Thanks to PR from [Yaroslav Pronin](https://github.com/proninyaroslav)
 * **Note**: the 0.5.0 release *does not* have the same as what was on the 0.5.0 pre-releases. Those pre-releases are more closely related to changes being done for the 10.0.0 pre-release of `flutter_local_notifications`. What was in the 0.5.0 pre-releases will be shifted to a 1.0.0 pre-release of `flutter_local_notifications_linux`
+
+## [0.5.0-dev.4]
+
+* **Breaking change** callbacks have now been reworked. `onDidReceiveNotificationResponse` is invoked only when the app is running. This works for when a user has selected a notification or notification action. This replaces the `onSelectNotification` callback that existed before
+* **Breaking change** the `NotificationAppLaunchDetails` has been updated to contain an instance `NotificationResponse` class with the `payload` belonging to the `NotificationResponse` class. This is to allow knowing more details about what caused the app to launch e.g. if a notification action was used to do so
+
+## [0.5.0-dev.3]
+
+* Includes changes from 0.4.2
+
+## [0.5.0-dev.2]
+
+* Added support for notification actions
+
+## [0.5.0-dev.1]
+
+* Bumped `flutter_local_notifications_platform_interface` dependency
 
 ## [0.4.2]